Code factorization.

+++ handler-project_update-tags.php
@@ -1,40 +1,49 @@
-<?php
+<?php session_start();
+if($_SESSION['username']){
 
-$project_id = strip_tags($_POST['project_id']);
+    if(isset($_POST['data_submit']) && !empty($_POST['data_submit'])){
 
+        $project_id = strip_tags($_POST['project_id']);
 
-require_once('db_connect.php');
+        require_once('db_connect.php');
 
-try {
+        try {
 
-    $sql = 'DELETE FROM `intermediary_tags-to-projects` WHERE `project_id` = :project_id;' ;
-    $query = $db->prepare($sql);
-    $db->beginTransaction();
-    $query->bindValue(':project_id', $project_id, PDO::PARAM_INT);
-    $query->execute();
+            $sql = 'DELETE FROM `intermediary_tags-to-projects` WHERE `project_id` = :project_id;' ;
+            $query = $db->prepare($sql);
+            $db->beginTransaction();
+            $query->bindValue(':project_id', $project_id, PDO::PARAM_INT);
+            $query->execute();
 
-    $sql = 'INSERT INTO `intermediary_tags-to-projects` (`project_id`, `tag_id`) VALUES (:project_id, :tag_id)';
-    $query = $db->prepare($sql);
-    foreach($_POST as $tag => $tag_id){
-        if ($tag != 'project_id') {
-                $query->bindValue(':project_id', $project_id, PDO::PARAM_INT);
-                $query->bindValue(':tag_id', $tag_id, PDO::PARAM_INT);
-                $query->execute();
+            $sql = 'INSERT INTO `intermediary_tags-to-projects` (`project_id`, `tag_id`) VALUES (:project_id, :tag_id)';
+            $query = $db->prepare($sql);
+            foreach($_POST as $tag => $tag_id){
+                if ($tag != 'project_id') {
+                        $query->bindValue(':project_id', $project_id, PDO::PARAM_INT);
+                        $query->bindValue(':tag_id', $tag_id, PDO::PARAM_INT);
+                        $query->execute();
+                }
+            }
+            
+            $db->commit();
+
+            $_SESSION['success'] = 'Tags have been updated. ';
+            header('Location: form-project_edit.php?project_id='.$project_id); 
+            
+        } catch (\Exception $e) {
+            if ($db->inTransaction()) {
+                $db->rollback();
+                // If we got here our two data updates are not in the database
+            }
+            throw $e;
         }
-    }
-    
-    $db->commit();
 
-    $_SESSION['success'] = 'Tags have been updated. ';
-    header('Location: form-project_edit.php?project_id='.$project_id); 
-    
-} 
-catch (\Exception $e) {
-    if ($db->inTransaction()) {
-        $db->rollback();
-        // If we got here our two data updates are not in the database
+        require_once('db_close.php');
+    } else {
+        $_SESSION['error'] = "Complete all fields.";
+        header('Location: view-backoffice_home.php');
     }
-    throw $e;
+} else {
+    $_SESSION['error'] = "Username or password are incorrect.";
+    header('Location: index.php');
 }
-
-
